DFS using user input provision in python
Implementation of Depth First Search algorithm with user input system, Where the user has the provision to 
dynamically create the graph of their choice without defining the graph inside the code.
In the pre-existing code for DFS in python, graph is defined within the code.  In this code, the graph can be created as needed when the code is being run.