BFS with user input provision
Implementation of Breadth First Search algorithm with user input system, Where the user has the provision to dynamically create the graph of their choice without defining the graph inside the code. This is useful because, it is very helpful to know how to create graph and tree like structures without using inbuilt functions.